chore(lib): moved one function to other file
